Factor out map-spawning logic into plugin
